Bug 821683/Bug 820317 - [Email] Confirmation overlay didn't sync with building-block. r=asuth. (cherry picked from commit fd1dce6e1b56a21a11b7c801408ecb57da1591b3) a=approval-gaia-v1 on bug 821683.
Conflict was from https://bugzilla.mozilla.org/show_bug.cgi?id=811858 not being
present on v1-train.

Conflicts:
	apps/email/js/compose-cards.js

+++ apps/email/js/setup-cards.js
@@ -712,23 +712,23 @@ SettingsAccountCard.prototype = {
 
   onDelete: function() {
     var account = this.account;
-    CustomDialog.show(
-      null,
-      mozL10n.get('settings-account-delete-prompt', { account: account.name }),
-      {
-        title: mozL10n.get('settings-account-delete-cancel'),
-        callback: function() {
-          CustomDialog.hide();
-        }
-      },
-      {
-        title: mozL10n.get('settings-account-delete-confirm'),
-        callback: function() {
+
+    var dialog = tngNodes['account-delete-confirm'].cloneNode(true);
+    var content = dialog.getElementsByTagName('p')[0];
+    content.textContent = mozL10n.get('settings-account-delete-prompt',
+                                      { account: account.name });
+    ConfirmDialog.show(dialog,
+      { // Confirm
+        id: 'account-delete-ok',
+        handler: function() {
           account.deleteAccount();
-          CustomDialog.hide();
           Cards.removeCardAndSuccessors(null, 'none');
           App.showMessageViewOrSetup();
         }
+      },
+      { // Cancel
+        id: 'account-delete-cancel',
+        handler: null
       }
     );
   },
